Revert "Uni code style"
This reverts commit 5c18308914be8a2bceb758b382f855fae1b83c8c.

+++ Assets/Source/World/Masks/SmoothJob.cs
@@ -3,13 +3,15 @@ using Unity.Collections;
 using Unity.Jobs;
 using static Unity.Mathematics.math;
 
-namespace Utopia.World.Masks {
+namespace Utopia.World.Masks
+{
 	/// <summary>
-	///     Parallel job to smooth the final part of the mask mesh vertices,
-	///     where the angles loop back around.
+	/// Parallel job to smooth the final part of the mask mesh vertices,
+	/// where the angles loop back around.
 	/// </summary>
 	[BurstCompile]
-	public struct SmoothJob : IJobParallelFor {
+	public struct SmoothJob : IJobParallelFor
+	{
 		// Input samples
 		public float startSample;
 		public float endSample;
@@ -18,12 +20,13 @@ namespace Utopia.World.Masks {
 		[WriteOnly] public NativeSlice<float> array;
 
 		/// <summary>
-		///     The amount of samples to replace.
+		/// The amount of samples to replace.
 		/// </summary>
 		public int smoothSamples;
 
-		public void Execute(int i) {
-			float position = i / (float)smoothSamples;
+		public void Execute(int i)
+		{
+			float position = (float) i / (float) smoothSamples;
 			array[i] = lerp(startSample, endSample, smoothstep(0, 1, position));
 		}
 	}
